fix(game): change list_external_games id column type from u32 to i64
cc_logic_game.id is BIGINT in MySQL, which is incompatible with Rust u32.
Changed return types and callers:
- list_external_games: Vec<(i64, String, String)>
- list_external_games_cached: Vec<(i64, String, String)>
- game_list builtin: entries Vec<(i64, String, Vec<String>)>

+++ crates/hiveweb/src/runtime/builtins/game_list.rs
@@ -54,7 +54,7 @@ async fn game_list_async_impl(
     };
 
     // 3. Merge: for each external game, collect all aliases (its own + internal supplements)
-    let mut entries: Vec<(u32, String, Vec<String>)> = Vec::new();
+    let mut entries: Vec<(i64, String, Vec<String>)> = Vec::new();
     for (id, name, ext_alias) in external {
         let mut aliases: Vec<String> = Vec::new();
